retain-original-class-as-interface: rename interface as RateLimiter to limit migration pain
